requirements.txt
torch
transformers==4.43.0
accelerate
pandas
numpy
matplotlib
seaborn
pyyaml
python-dotenv
